moved BitvectorDomain into its own source file.
