+++ cwe_checker_rs/src/analysis/pointer_inference/object.rs
@@ -1,6 +1,6 @@
 use super::data::*;
 use super::identifier::AbstractIdentifier;
-use crate::analysis::abstract_domain::*;
+use crate::abstract_domain::*;
 use crate::analysis::mem_region::MemRegion;
 use crate::bil::Bitvector;
 use crate::prelude::*;
